+++ FireworkRNSDKSample/src/screens/paramList/TopParamList.ts
@@ -1,0 +1,7 @@
+export type TopParamList = {
+  Cart: undefined;
+  Checkout: undefined;
+  LinkContent: {
+    url: string;
+  };
+};